Removed save functionality from player script to separate script and node.
