+++ SaveData/SaveData.json
@@ -1,1 +1,1 @@
-{"Health":37.5000000000001}
+{"Health":100,"Player":"{\"Position\":{\"X\":0.0,\"Y\":0.0},\"Rotation\":0.0,\"RotationDegrees\":0.0,\"Scale\":{\"X\":1.0,\"Y\":1.0},\"Skew\":0.0,\"Transform\":{\"X\":{\"X\":1.0,\"Y\":0.0},\"Y\":{\"X\":0.0,\"Y\":1.0},\"Origin\":{\"X\":0.0,\"Y\":0.0},\"Rotation\":0.0,\"Scale\":{\"X\":1.0,\"Y\":1.0},\"Skew\":0.0},\"GlobalPosition\":{\"X\":0.0,\"Y\":0.0},\"GlobalRotation\":0.0,\"GlobalRotationDegrees\":0.0,\"GlobalScale\":{\"X\":1.0,\"Y\":1.0},\"GlobalSkew\":0.0,\"GlobalTransform\":{\"X\":{\"X\":1.0,\"Y\":0.0},\"Y\":{\"X\":0.0,\"Y\":1.0},\"Origin\":{\"X\":0.0,\"Y\":0.0},\"Rotation\":0.0,\"Scale\":{\"X\":1.0,\"Y\":1.0},\"Skew\":0.0},\"Visible\":true,\"Modulate\":{\"R\":1.0,\"G\":1.0,\"B\":1.0,\"A\":1.0,\"R8\":255,\"G8\":255,\"B8\":255,\"A8\":255,\"H\":0.0,\"S\":0.0,\"V\":1.0,\"Luminance\":1.0},\"SelfModulate\":{\"R\":1.0,\"G\":1.0,\"B\":1.0,\"A\":1.0,\"R8\":255,\"G8\":255,\"B8\":255,\"A8\":255,\"H\":0.0,\"S\":0.0,\"V\":1.0,\"Luminance\":1.0},\"ShowBehindParent\":false,\"TopLevel\":false,\"ClipChildren\":0,\"LightMask\":1,\"VisibilityLayer\":1,\"ZIndex\":0,\"ZAsRelative\":true,\"YSortEnabled\":false,\"TextureFilter\":0,\"TextureRepeat\":0,\"Material\":null,\"UseParentMaterial\":false,\"_ImportPath\":{\"IsEmpty\":true},\"Name\":{\"IsEmpty\":false},\"UniqueNameInOwner\":false,\"SceneFilePath\":\"res://Scenes/Helper/SaveNode.tscn\",\"Owner\":{\"_ImportPath\":{\"IsEmpty\":true},\"Name\":{\"IsEmpty\":false},\"UniqueNameInOwner\":false,\"SceneFilePath\":\"res://Main.tscn\",\"Owner\":null,\"Multiplayer\":{\"RootPath\":{\"IsEmpty\":false},\"AuthCallback\":{\"Target\":null,\"Method\":{\"IsEmpty\":true},\"Delegate\":null,\"Trampoline\":{\"value\":0}},\"AuthTimeout\":3.0,\"AllowObjectDecoding\":false,\"RefuseNewConnections\":false,\"ServerRelay\":true,\"MaxSyncPacketSize\":1350,\"MaxDeltaPacketSize\":65535,\"MultiplayerPeer\":{\"RefuseNewConnections\":false,\"TransferMode\":2,\"TransferChannel\":0,\"EncodeBufferMaxSize\":8388608,\"NativeInstance\":{\"value\":2072992993296}},\"NativeInstance\":{\"value\":2073074414128}},\"ProcessMode\":0,\"ProcessPriority\":0,\"ProcessPhysicsPriority\":0,\"ProcessThreadGroup\":0,\"ProcessThreadGroupOrder\":0,\"ProcessThreadMessages\":0,\"EditorDescription\":\"\",\"NativeInstance\":{\"value\":2073074481296}},\"Multiplayer\":{\"RootPath\":{\"IsEmpty\":false},\"AuthCallback\":{\"Target\":null,\"Method\":{\"IsEmpty\":true},\"Delegate\":null,\"Trampoline\":{\"value\":0}},\"AuthTimeout\":3.0,\"AllowObjectDecoding\":false,\"RefuseNewConnections\":false,\"ServerRelay\":true,\"MaxSyncPacketSize\":1350,\"MaxDeltaPacketSize\":65535,\"MultiplayerPeer\":{\"RefuseNewConnections\":false,\"TransferMode\":2,\"TransferChannel\":0,\"EncodeBufferMaxSize\":8388608,\"NativeInstance\":{\"value\":2072992993296}},\"NativeInstance\":{\"value\":2073074414128}},\"ProcessMode\":0,\"ProcessPriority\":0,\"ProcessPhysicsPriority\":0,\"ProcessThreadGroup\":0,\"ProcessThreadGroupOrder\":0,\"ProcessThreadMessages\":0,\"EditorDescription\":\"\",\"NativeInstance\":{\"value\":2073074475408}}"}
