+++ Scenes/Helper/SaveHelper.cs
@@ -1,15 +1,10 @@
 using Godot;
+using Newtonsoft.Json;
 using System;
+using TheLoneLanternProject.Constants;
 
 namespace TheLoneLanternProject.Scenes.SaveHelper;
 public partial class SaveHelper : Node
 {
-    public static Godot.Collections.Dictionary<string, Variant> Save()
-    {
-        return new Godot.Collections.Dictionary<string, Variant>()
-        {
-            {"CurrentScene","2"},
-            //{"Health": },
-        };
-    }
+    
 }
