+++ Scenes/Helper/SaveNode.tscn
@@ -1,0 +1,6 @@
+[gd_scene load_steps=2 format=3 uid="uid://imn0en4fbde3"]
+
+[ext_resource type="Script" path="res://Scenes/Helper/SaveNode.cs" id="1_lruhr"]
+
+[node name="SaveNode" type="Node2D"]
+script = ExtResource("1_lruhr")
