+++ The Lone Lantern Project.csproj
@@ -7,4 +7,7 @@
   <ItemGroup>
     <Folder Include="Areas\" />
   </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+  </ItemGroup>
 </Project>
